[FIX] l10n_ar_tax: Prevent duplicate tax withholdings when splitting payments
Ensure that tax withholdings are not recalculated when a payment is split, particularly when changing the journal in the second payment. To achieve this, added a context flag to detect when the payment is being created using the "Create and New" button.
This prevents the system from incorrectly applying withholdings again in the new payment.

closes ingadhoc/odoo-argentina#1059

Related: ingadhoc/account-payment#628
Signed-off-by: Filoquin adhoc <maq@adhoc.com.ar>